+++ server/config/toml.go
@@ -25,8 +25,8 @@ evm-chain-id = {{ .EVM.EVMChainID }}
 # MinTip defines the minimum priority fee for the mempool.
 min-tip = {{ .EVM.MinTip }}
 
-# GethMetricsAddr defines the addr to bind the geth metrics server to. Default 127.0.0.1:8080.
-geth-metrics-addr = {{ .EVM.GethMetricsAddr }}
+# GethMetricsAddress defines the addr to bind the geth metrics server to. Default 127.0.0.1:8100.
+geth-metrics-address = "{{ .EVM.GethMetricsAddress }}"
 
 ###############################################################################
 ###                           JSON RPC Configuration                        ###